Internal: Move element related base classes into the elements folder [ED-17488]
