+++ modules/atomic-widgets/utils/utils.php
@@ -3,8 +3,8 @@
 namespace Elementor\Modules\AtomicWidgets\Utils;
 
 use Elementor\Core\Base\Document;
-use Elementor\Modules\AtomicWidgets\Elements\Atomic_Element_Base;
-use Elementor\Modules\AtomicWidgets\Elements\Atomic_Widget_Base;
+use Elementor\Modules\AtomicWidgets\Elements\Base\Atomic_Element_Base;
+use Elementor\Modules\AtomicWidgets\Elements\Base\Atomic_Widget_Base;
 use Elementor\Plugin;
 
 if ( ! defined( 'ABSPATH' ) ) {
