+++ tests/phpunit/elementor/modules/atomic-widgets/styles/test-atomic-widgets-base-styles.php
@@ -5,7 +5,7 @@ namespace Elementor\Testing\Modules\AtomicWidgets\Styles;
 use Elementor\Core\Files\CSS\Post;
 use Elementor\Elements_Manager;
 use Elementor\Modules\AtomicWidgets\Styles\CacheValidity\Cache_Validity;
-use Elementor\Modules\AtomicWidgets\Elements\Atomic_Widget_Base;
+use Elementor\Modules\AtomicWidgets\Elements\Base\Atomic_Widget_Base;
 use Elementor\Modules\AtomicWidgets\PropTypes\Color_Prop_Type;
 use Elementor\Modules\AtomicWidgets\PropTypes\Primitives\String_Prop_Type;
 use Elementor\Modules\AtomicWidgets\Styles\Atomic_Styles_Manager;
