+++ tests/phpunit/elementor/modules/atomic-widgets/template-renderer/test-template-renderer.php
@@ -2,8 +2,8 @@
 
 namespace Elementor\Testing\Modules\AtomicWidgets\TemplateRenderer;
 
-use Elementor\Modules\AtomicWidgets\Elements\Atomic_Widget_Base;
-use Elementor\Modules\AtomicWidgets\Elements\Has_Template;
+use Elementor\Modules\AtomicWidgets\Elements\Base\Atomic_Widget_Base;
+use Elementor\Modules\AtomicWidgets\Elements\Base\Has_Template;
 use Elementor\Modules\AtomicWidgets\Styles\Style_Definition;
 use Elementor\Modules\AtomicWidgets\Elements\TemplateRenderer\Template_Renderer;
 use ElementorEditorTesting\Elementor_Test_Base;
